Write to cognate db on fewer edits
This now checks the previous and current state of the
page in more detail and will only update the cognate
DB in the cases we are sure we want to make a change.

These are:
 - New page that is not a redirect
 - Page that was a redirect that now is not
 - Page that was not a redirect that now is

Tests has also been added for these cases & all cases between
to assert no writing to the db happens.

Bug: T156239
Change-Id: I5bbd516f944798253ca78c7c1eb2c19c5c894b37