CognateUpdater, use default DB for __construct call
This makes the CognateUpdater newForCognateDB method
use the regular DB used by the main updater while
DatabaseUpdater::__construct is called.
The DB is then updated to be the cognate DB.

This is needed as during __construct the LoadExtensionSchemaUpdates
hooks are run, which can call any number of public methods on
the updater instance.
For example, Flow calls the updateRowExists method, which will
fail on the Cognate DB.

The doUpdates method in CognateUpdater has been removed
as it is never called as CognateUpdater itself is not
instansiated.

Depends-On: I57b1fbbe2f85a74d56c49edef012a185d32ceb1a
Change-Id: Iadc36d380e71b73e531b29c46fe6ceef9cf21f8b