Fix selectSitesForPage to use getReadConnectionRef
This select is made after edits that cause the links
to change for pages (such as creations, moves and deletes)

Prior to Iff20a22f9f2bc7ceefd6defc0ed9a494a6fe62c0 this
used a read connection and the patch accidently changed this
method to using a write connection.

Probably not the cause of the attached bug but could be a factor.

Bug: T164407
Change-Id: Ia1ddc54d800ab1695f0909b81849670ac35e4834