Add @covers tags to CognateIntegrationTest
I realized this test is actually testing the code in CognateHooks, which
was entirely uncovered by now:
https://doc.wikimedia.org/cover-extensions/Cognate/src/

As it is an integration test, it is testing the entire chain:
* MediaWiki triggers one of the hook handlers in CognateHooks.
* But that's only a little bit of glue code. The actual hook handlers
  are in CognatePageHookHandler.
* That again calls save/delete methods in CognateRepo.
* Which calls save/delete code in CognateStore.
* CognateIntegrationTest then checks if the store succeeded.

A lot of this was not covered by unit tests before. I think it is fine
for this code to not necessarily have separate unit tests, but be
covered by this integration test instead.

Change-Id: Ia76ca0be6183c88fd5451d76c63837db08ace098