Only pass required arguments to hook handler functions
This codebase uses an extra level of indirection for the hook handler
functions. This allows to greatly simplify the code to only require
arguments it actually needs and uses.

This greatly simplifies the test setup. All the dummy stuff that does
not do anything can be removed now.

I'm also removing the strict type hints from all arguments that are
unused by this code. This makes this code less brittle when changes are
made to parameters this code does not care about.

This patch also removes the default `return true`. This doesn't do
anything, as it is the default anyway.

Change-Id: I25d01c9b87ce800ba6a8ead6ca3024cf9ad65e96