Fix incompatible CacheUpdateJob signature
All Job subclasses must have the same signature, with a $params array
as the second parameter. Jobs are serialized. The constructor is
expected to be exactly like this during deserialization. Luckily PHP
ignores extra parameters that are passed to a function that doesn't
expect them. That's why this was never a problem. But it also means
that extra parameters the parent class might need previously got lost.

Change-Id: I4587b4d652ad98e93f2546c09ddc957555585719