Remove redundant property from LocalJobSubmitJob
The parent constructor stores all parameters in $this->params. This is
also what is serialized. The extra private property does not really hurt
(it's just a reference, not really consuming memory). But redundant
anyway.

This patch also moves the array_unique() call. The effect is that the
array will typically not contain a duplicate any more. Doing this outside
of the job does have the advantage that this duplicate is not serialized
and stored in the database when the job is serialized.

Change-Id: I12409807288b3c2e34d29631b5f842571fb11812