Exit undelete hook early if revision not found
Revision::newFromId() can return null when it is unable to
find the Revision. This patch adds a check for this case and bails
early.

Note:
- this patch continues the existing pattern of using Revision
rather than trying to migrate to RevisionStore and RevisionRecord
and adds to the use of the testing access wrapper.
- this patch does not attempt to add return types (which could have
helped catch the bug) since the README specifies php < 7

Bug: T245869
Change-Id: I7653162bf758c766e80fb647980e691f78e69eaf