Do not extend DatabaseUpdater, copy its functionality instead
There were only a couple of lines of useful code in DatabaseUpdater, and
more lines than that in workarounds for problems with using it. We need
to stop using it now since DatabaseUpdater is changing in an
incompatible way.

The --noschema and --schema options to update.php are broken, but they
were broken before this change, since the CognateUpdater does not have
access to the parameters to doUpdate() passed down from the maintenance
script, regardless of whether it extends DatabaseUpdater.

Bug: T249584
Bug: T249732
Change-Id: I3c88513bdee7913f3674d81677394bd022ea4d28