Lazy push CacheInvalidator job
That job is being queued in all edits and they are being pushed directly.
This means in every edit, the appserver opens a request, communicate to
eventgate and then closes it, adding around 40ms to the time to save the edit.

You can see it in flamegraph of Pre-send deferred updates:
https://performance.wikimedia.org/arclamp/svgs/daily/2021-09-30.excimer-wall.all.fn-PreSend.svgz

It's not an urgent job and can be pushed later through deferred update
instead. The important thing is that it gets batched through all other
jobs being pushed.

Bug: T292048
Bug: T255502
Change-Id: I90967164656b6351d93230389b57643e2cdbc694