Migrate to virtual domain mapping
This way Cognate will use the defined virtual domain,
if available. If the virtual domain is not defined,
the local database will be used as a fallback, thus
no further configuration will be needed for such a
set up.

We don't need to add a migration path here, as this
extension is only used on WMF wikis[0], and we
already have the virtual domain mapping set up.

[0]: https://wikiapiary.com/wiki/Extension:Cognate

Bug: T348526
Change-Id: Ib6f47bbd8614fe187981ce062f192817c80805ab