Migrate addWiki populateCognateSites to an installer task
addWiki.php in WikimediaMaintenance runs the populateCognateSites
maintenance script, if Cognate is enabled.

It seems like overkill. All we need is to insert the local interwiki
prefix and the wiki ID of the wiki being created, there's no need to
rebuild the entire table. There's no need for a dependency on the sites
table.

So, now that I am migrating addWiki.php to core, add an installer task
that inserts the one new row into the table, using the configured
$wgLocalInterwikis[0], which is always set to $lang in
InitialiseSettings.php.

Remove the unconventional check for CLI mode since I can't guarantee
that there won't be a web interface to the preconfigured installer in
future.

Bug: T352113
Change-Id: Id3c31ac4615d296270a7063dfd03afd37986ef6d